Fixed Segue for viewing data from recorder. Something seems to be broken however, I think the load data is being called too soon.

+++ ExpressiveTouch/ForceDetector.swift
@@ -23,7 +23,7 @@ class ForceDetector {
         var gForce = 0.0
         
         if (accCache.count() > 0) {
-            var index = accCache.count() - 1
+            var index = UInt(accCache.count() - 1)
             var x = accCache.get(index).x / 9.81
             var y = accCache.get(index).y / 9.81
             var z = accCache.get(index).z / 9.81
